+++ src/Output.js
@@ -270,15 +270,12 @@ export class Output extends EventEmitter {
   /**
    * Sends a MIDI [system exclusive]{@link
     * https://www.midi.org/specifications-old/item/table-4-universal-system-exclusive-messages}
-   * (*sysex*) message. The `data` parameter should only contain the actual data of the message.
-   * When sending out the actual MIDI message, WebMidi.js will automatically prepend the data with
-   * the *sysex byte* (`0xF0`) and the manufacturer ID byte(s). It will also automatically terminate
+   * (*sysex*) message. The `data` parameter should only contain the data of the message. When
+   * sending out the actual MIDI message, WebMidi.js will automatically prepend the data with the
+   * *sysex byte* (`0xF0`) and the manufacturer ID byte(s). It will also automatically terminate
    * the message with the *sysex end byte* (`0xF7`).
    *
-   * The data can be an array of unsigned integers or a `Uint8Array` object.
-   *
-   *
-   *
+   * The data can be an array of unsigned integers (0-127) or a `Uint8Array` object.
    *
    * To use the `sendSysex()` method, system exclusive message support must have been enabled. To
    * do so, you must set the `sysex` option to `true` when calling `WebMidi.enable()`:
@@ -326,8 +323,8 @@ export class Output extends EventEmitter {
    * [Manufacturer ID Numbers](https://www.midi.org/specifications-old/item/manufacturer-id-numbers)
    * .
    *
-   * @param [data=number[]] {Array} An array of unsigned integers between 0 and 127. This is the
-   * data you wish to transfer.
+   * @param [data=number[]] {number[]|Uint8Array} A Uint8Array or an array of unsigned integers
+   * between 0 and 127. This is the data you wish to transfer.
    *
    * @param {Object} [options={}]
    *
@@ -346,18 +343,20 @@ export class Output extends EventEmitter {
    */
   sendSysex(manufacturer, data, options = {}) {
 
-    // Merging Uint8Arrays
-    // var arrayOne = new Uint8Array([2,4,8]);
-    // var arrayTwo = new Uint8Array([16,32,64]);
-    //
-    // var mergedArray = new Uint8Array(arrayOne.length + arrayTwo.length);
-    // mergedArray.set(arrayOne);
-    // mergedArray.set(arrayTwo, arrayOne.length);
-
     manufacturer = [].concat(manufacturer);
 
-    data = manufacturer.concat(data, WebMidi.MIDI_SYSTEM_MESSAGES.sysexend);
-    this.send([WebMidi.MIDI_SYSTEM_MESSAGES.sysex].concat(data), {time: options.time});
+    // Check if data is Uint8Array
+    if (data instanceof Uint8Array) {
+      const merged = new Uint8Array(1 + manufacturer.length + data.length + 1);
+      merged[0] = WebMidi.MIDI_SYSTEM_MESSAGES.sysex;
+      merged.set(Uint8Array.from(manufacturer), 1);
+      merged.set(data, 1 + manufacturer.length);
+      merged[merged.length - 1] = WebMidi.MIDI_SYSTEM_MESSAGES.sysexend;
+      this.send(merged, {time: options.time});
+    } else {
+      const merged = manufacturer.concat(data, WebMidi.MIDI_SYSTEM_MESSAGES.sysexend);
+      this.send([WebMidi.MIDI_SYSTEM_MESSAGES.sysex].concat(merged), {time: options.time});
+    }
 
     return this;